refactor(maps): move components to src folder
